rag-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.40.0
httpx>=0.27.0,<0.28.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
chromadb==0.4.22
sentence-transformers==2.2.2
